Move usestore to new folder
